Only use ocr if file has less than five pages

+++ app/Jobs/ConvertPDF.php
@@ -71,15 +71,20 @@ class ConvertPDF implements ShouldQueue
                     for($x = 0; $x < $numPages; $x++) {
                         $num_padded = sprintf("%02d", $x);
                         $im->setIteratorIndex($x);
-                        $im->writeImage($savePath . 'main_' . $num_padded . '.png');
+                        if($numPages < 5) {
+                            $im->writeImage($savePath . 'main_' . $num_padded . '.png');
+                        }
                         $im->thumbnailImage(650, 0);
                         $d = $im->getImageGeometry();
                         $im->setImageFormat('png');
                         $im->writeImage($savePath . 'image_' . $num_padded . '.png');
-                        $ocrData = (new TesseractOCR($savePath . 'main_' . $num_padded . '.png'))->run();
-                        $pdfText['page'.$num_padded] = $ocrData;
-                        File::delete($savePath . 'main_' . $num_padded . '.png');
-
+                        if($numPages < 5) {
+                            $ocrData = (new TesseractOCR($savePath . 'main_' . $num_padded . '.png'))->run();
+                            $pdfText['page'.$num_padded] = $ocrData;
+                        }
+                        if($numPages < 5) {
+                            File::delete($savePath . 'main_' . $num_padded . '.png');
+                        }
                         $files[$x]['width'] = 650;
                         $files[$x]['height'] = $d['height'];
                         $files[$x]['file'] = 'image_' . $num_padded . '.png';
@@ -115,9 +120,9 @@ class ConvertPDF implements ShouldQueue
 
                         $this->entry->files = json_encode($files);
                         $this->entry->save();
-
-                        TextLogic::create($this->project, json_encode($pdfText));
-
+                        if($numPages < 5) {
+                            TextLogic::create($this->project, json_encode($pdfText));
+                        }
 
 
                         if($orderVals->order->notify_users) {
